fix: router simplified
